Fix: Rename Android samples package
