+++ sentry-samples/sentry-samples-android/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="io.sentry.sample">
+    package="io.sentry.samples">
 
     <!--  extra recommended permission, we'd be able to check network status-->
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
