+++ sentry-samples/sentry-samples-android/src/main/cpp/native-sample.cpp
@@ -6,13 +6,13 @@
 
 extern "C" {
 
-JNIEXPORT void JNICALL Java_io_sentry_sample_NativeSample_crash(JNIEnv *env, jclass cls) {
+JNIEXPORT void JNICALL Java_io_sentry_samples_NativeSample_crash(JNIEnv *env, jclass cls) {
     __android_log_print(ANDROID_LOG_WARN, TAG, "About to crash.");
     char *ptr = 0;
     *ptr += 1;
 }
 
-JNIEXPORT void JNICALL Java_io_sentry_sample_NativeSample_message(JNIEnv *env, jclass cls) {
+JNIEXPORT void JNICALL Java_io_sentry_samples_NativeSample_message(JNIEnv *env, jclass cls) {
     __android_log_print(ANDROID_LOG_WARN, TAG, "Sending message.");
     sentry_value_t event = sentry_value_new_message_event(
             /*   level */ SENTRY_LEVEL_INFO,
