+++ sentry-samples/sentry-samples-android/src/main/java/io/sentry/samples/MyApplication.java
@@ -1,4 +1,4 @@
-package io.sentry.sample;
+package io.sentry.samples;
 
 import android.app.Application;
 import android.os.StrictMode;
